ロジックとレイアウトを分割
